requirements.txt
# AgentQA Backend Dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.6.5
langgraph>=0.2.0
langchain-core>=0.2.0
httpx>=0.27.0
requests>=2.32.3
reportlab>=4.2.2
pytest>=8.2.2
pytest-cov>=5.0.0
coverage>=7.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
google-generativeai>=0.7.0
